ez: do not trust local commands to properly reflect
As we just discovered, Google's sendgmr is not properly compatible with
sendmail flags, and will ignore addresses passed via the CLI, instead
always using the ones from the headers.

This is opposite to what --reflect is designed to do, so bail out with a
loud error, but give a way to override it via a custom variable.

Reported-by: Nick Desaulniers <ndesaulniers@google.com>
Link: https://bugzilla.kernel.org/show_bug.cgi?id=217332
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>