b4: Allow prep new branch while on a b4 managed branch
Creating a new branch from a specified fork point should be allowed when
sitting on a b4 managed branch.  For example this fails:

	$ b4 prep -n foo -f master
	Created new branch b4/foo
	Created the default cover letter, you can edit with --edit-cover.

	$ b4 prep -n bar -f master
	CRITICAL: This appears to already be a b4-prep managed branch.

Allow a specified fork point to be used regardless of the b4 state of
the current branch.

Make -n and -e options mutually exclusive early on to clarify the
difference between them in practical usage.

With this change the following works as expected.

	$ git co b4/allow-new-w-fork-point
	Switched to branch 'b4/allow-new-w-fork-point'

	$ b4 prep -n bar -f master
	Created new branch b4/bar
	Created the default cover letter, you can edit with --edit-cover.

	$ b4 prep -n foobar -e master -f master
	CRITICAL: -n NEW_SERIES_NAME and -e [ENROLL_BASE] can not be used together.

Cc: tools@linux.kernel.org
Cc: konstantin@linuxfoundation.org
Signed-off-by: Ira Weiny <ira.weiny@intel.com>
Link: https://msgid.link/20230517-allow-new-w-fork-point-v1-1-22d2cb938e16@intel.com
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>