b4: Fix envelopeSender handling
The manpage for git-send-email states that:

  Specify the envelope sender used to send the emails. This is useful if
  your default address is not the address that is subscribed to a list. In
  order to use the From address, set the value to "auto". If you use the
  sendmail binary, you must have suitable privileges for the -f parameter.
  Default is the value of the sendemail.envelopeSender configuration
  variable; if that is unspecified, choosing the envelope sender is left to
  your MTA.

However, b4 just reuses the value of sendemail.envelopeSender and passes it
down directly to the sendmail binary if relevant, which means that if we
were using the value "auto", we now have sendemail called with -f auto
instead of the proper mail address, confusing sendmail.

We can fix this by interpreting the auto value as suggested in the manpage.

Signed-off-by: Maxime Ripard <mripard@kernel.org>
Link: https://msgid.link/20230602152554.2046764-1-mripard@kernel.org
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>