ez: only include base-branch when using non-default strategy
When we are using the default "commit" strategy, we do not really use
the base-branch info for anything. Keeping it in the tracking json for
historical reasons actually confuses people, because the base-branch is
not updated when operations such as "git rebase" are performed.

For this reason, only include "base-branch" when we're using a strategy
other than "commit".

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>