ez: Add --use-web-endpoint arg to force web endpoint
Make it easier to prefer going through the web endpoint by adding an
argument that will force this path. This is for users that want to use
B4 for preparing and sharing patchsets over mail and need an argument to
activate the web endpoint whenever needed, but want to default to smtp
server for the rest.

Signed-off-by: Joel Granados <j.granados@samsung.com>
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>