mbox: retrieve code review trailers matching the patch-id
Initial implementation to retrieve code-review trailers not just sent as
follow-ups in the current thread, but also sent to any other thread with
the same patch.

This helps in situations where a trailer is sent to a previous revision
of the series, for example. Similarly, if someone sends a large series
and then breaks it up into smaller ones, we'll still retrieve any
code-review trailers as long as the patch-id matches.

The downside is that we're sending an additional remote query to lore,
but it's tolerably fast in my tests.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>