ez: implement send --dry-run-to
Sometimes it is useful to be able to send the series out to someone for
a quick one-over review (e.g. to your mentor, boss, colleague) before
actually sending it out to the actual maintainers.

It is now possible to do so with:

    b4 send --dry-run-to addr@example.com

This will not trigger a tag-and-reroll and will include a DRYRUN into
the patch prefixes to clearly indicate that it's not the final
submission that should be used for review.

Suggested-by: Marcos Paulo de Souza <mpdesouza@suse.com>
Link: https://lore.kernel.org/f5c38763b4d42cfafdaac24d83ec18b81dfc073a.camel@suse.com
Closes: https://bugzilla.kernel.org/show_bug.cgi?id=218302
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>