b4: add --config option to set config options on the command-line
It's useful to be able to set and override configuration options on the
command-line. Add the --config and -c options to do this. This is
similar to the git -c option.

The basic form is 'b4 -c name=value'. Mimic git -c option for the 'b4 -c
name' and 'b4 -c name=' forms to set the value to 'true' and the empty
string, respectively. The name is also the same section.key dotted
format as in git -c. For example, 'b4 -c
b4.midmask=https://some.host/%s'.

The argparse action class is handy for special parsing. For details, see
[1] and [2]. This stores a dict of the config options set on the command
line to cmdargs.config, mapping the full dotted keys to values. Relevant
parts for main, user and sendemail configs will be set using
_cmdline_config_override().

[1] https://docs.python.org/3/library/argparse.html#action
[2] https://docs.python.org/3/library/argparse.html#action-classes

Signed-off-by: Jani Nikula <jani.nikula@intel.com>
Link: https://msgid.link/20240118202544.2019009-3-jani.nikula@intel.com
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>