mbox: add --single-message mode
This change adds ability to retrieve a single message and ignore any
follow-ups. When in that mode, we will additionally check if there are
any patches attached to that message, in case someone sent a single
message with a bunch of patches attached as a test series.

This is kinda hacky, but it solves a real problem, so it's worth the
added complexity.

Suggested-by: Steven Rostedt <rostedt@goodmis.org>
Link: https://lore.kernel.org/CAHk-=wh97AkwaOkXoBgf0z8EP88ePffLnTcmmQXcY+AhFaFrnA@mail.gmail.com
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>