ez: fall back to regular merge-base if --fork-point mode fails
If the 'git merge-base --fork-point' invocation in
ez::get_base_forkpoint fails and returns nothing, we abort, saying
basebranch and mybranch have no common ancestors.

Depending on how the current branch was created, and the state of the
branch's reflog, it is possible for this invocation to fail even if the
current branch and the chosen base branch do share a common ancestor,
for reasons expplained in git-merge-base(1) [1].

Support that situation better by adding a second invocation, using the
regular merge-base syntax 'git merge-base mybranch basebranch', and only
fail if that second invocation also returns nothing.

[1] https://git-scm.com/docs/git-merge-base#_discussion_on_fork_point_mode

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>
Link: https://msgid.link/20240210-ez-merge-base-no-fork-point-v1-1-b3f82cdd3bee@gmail.com
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>