ez: allow inserting a range-diff in the cover-letter
Some projects recommend inserting in the cover letter a range-diff
against the previous version when sending subsequent versions of a
series.

Support this convention by recognizing the token '${range_diff}' in the
cover letter template, and automatically invoking 'git range-diff'
against the previous revision when creating the cover letter.

As such, adjust ez.py::compare to allow returning the range-diff output
instead of 'execvp'-ing the command.

Since the range-diff invocation might take some time, only perform it if
the corresponding token is found in the cover letter template.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>