Initial support for retrieving dependencies
This is the first effort to automatically retrieve dependencies, for now
just those specified as prerequisite-patch-id entries. This is not very
widely used "in the wild," so I don't expect this to be very useful. I'm
hoping that with the increased use of b4-prep managed series and the
introduction of "prerequisite-change-id" in the future, this will become
more useful.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>