ez: initial support for chained series
When we are inside a prep-tracked branch and then try to create a brand
new tracked branch, warn the user that it will automatically mark the
new series as depending on the existing one.

Creates a new tracking var called "prerequisites" that is a list of
free-form strings in the format:

    change-id: some-change-id:vX
    patch-id: hex-of-a-patch-id

This will probably see some refinement, so this is potentially not the
final form it will take.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>