ez: check that all dependencies can be applied
Introduces prerequisite-base-commit, which is how we are supposed to
indicate where in the tree the dependencies + series patches should be
applied. It doesn't really make sense to me to use base-commit for it,
but I will check on the git list if I'm working with the correct
assumption in mind.

More testing and documentation is needed before this is considered
complete.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>