Allow defining multiple check commands
This lets us specify multiple commands that can run perpatch checks, for
example a checkpatch.pl plus some other custom command. We still expect
these commands to return a single line per check and contain the words
'WARNING' or 'ERROR' in order to properly work, so this isn't actually
that customizable.

That said, I feel focusing too much on local checks is the wrong
direction here, because this really needs to be done using a system like
patchwork or others.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>