Improve auto-noparenting logic to require a diff/diffstat
This is a partial fix for the reported problem -- we still need to be
able to retrieve the full thread in mbox mode, even if we're pointing at
a standalone patch in the middle of a thread. The auto-noparenting
should be specific to the am mode. That said, with this change, this
will likely behave as expected in the vast majority of situations.

Reported-by: Mark Brown <broonie@kernel.org>
Closes: https://bugzilla.kernel.org/show_bug.cgi?id=218933
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>