Report ecode and stderr from check commands
The ecode and stderr from check commands gets silently ignored. This
means if the check failed for some unrelated reason, it will appear as a
success to the patch author.

Treat all output from stderr as failure lines, as it's expected normal
findings would go to stdout.

Treat an error from the ecode as meaningful only if stdout and stderr
were blank, as checkpatch normally returns an ecode error if there are
any findings.

Signed-off-by: Brandon Maier <brandon.maier@collins.com>
Link: https://patch.msgid.link/20240626-report-check-cmd-stderr-v1-1-50d97e394371@collins.com
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>