Support "valueless truth" in configuration parsing
With a configuration item somewhere (like ~/.gitconfig) like this:

    [sendemail]
        suppressfrom

"b4" barfs while parsing the configuration file.

The code assumes that "git config -z" output, after splitting at
NUL, would always be a LF separated variable and value, but that
does not hold for such a variable defined with only its name without
"= <value>", which means that variable is given the value "true".

Signed-off-by: Junio C Hamano <gitster@pobox.com>
Link: https://patch.msgid.link/xmqq34o6zrmj.fsf@gitster.g
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>