Differentiate between checkcmd error and checkmd patch fail
If the check command fails for a non-patch related reason, if for
example the b4 config includes a wrong flag for the checkpatch command
or has an invalid command in general, it is not possible to disambiguate
whether the failure is with the patch or with the checkpatch invocation
instead.

This commit checks the stderr output of the check command, and prints it
if it's not empty. Furthermore, it does not cache the output since the
check was never essentially performed.

Signed-off-by: Manos Pitsidianakis <manos.pitsidianakis@linaro.org>