ez: use a temporary workdir when preparing sent tags
Instead of operating on the current checkout, use a temporary worktree
and a sparse checkout when preparing the tree to tag as a sent version.

Doesn't fix the bug reported, but will do a better job not leaving the
tree in a messed up stage.

Link: https://bugzilla.kernel.org/show_bug.cgi?id=219230
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>
(cherry picked from commit 7be4fe35d52daac57b8c6df7c3cf76218add771c)