change the manpage section from 5 to 1
Manpages in category 5 are according to 'man man' docs for "File formats
and conventions, e.g. /etc/passwd", which does not fit the contents of
the manpage in question. Instead move it to section 1, which contains
docs for "Executable programs or shell commands".

Since this also requires a small code change in the docs themselves also
regenerate the rendered version of the manpage.

Signed-off-by: Christian Heusel <christian@heusel.eu>
Link: https://patch.msgid.link/20240908202331.386220-1-christian@heusel.eu
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>