src/tests: fix trailer test hiding parser errors
The trailer parsing test compares the values of the b4.LoreTrailer
object created from the test input against a b4.LoreTrailer object
created from our expected test input.

This means if there is an error in how b4.LoreTrailer parses input, it
won't be detected. This was discovered from a reported bug with parsing
email trailers with embedded ','.

Bug: https://github.com/mricon/b4/issues/50
Link: https://patch.msgid.link/20241027-trailer-special-chars-v2-2-e963f78b3ef8@gmail.com
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>