b4: handle trailers with improper formatting
A patch trailer with an improperly formatted email parses to an empty
string. For example see the linked github issue, where a trailer's
display name contains an unquoted comma.

As patch trailers are often hand written this can happen frequently. Try
to gracefully handle this case by checking if an email parses correctly,
and if it doesn't fallback to displaying it verbatim.

Fixes: https://github.com/mricon/b4/issues/50
Signed-off-by: Brandon Maier <brandon.maier@gmail.com>
Link: https://patch.msgid.link/20241027-trailer-special-chars-v2-3-e963f78b3ef8@gmail.com
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>