ez: support shallow threads in "b4.send-same-thread"
When setting "b4.send-same-thread" to a truish value, B4 will send
revised versions of a patch series in the same thread. This is done by
setting the "In-Reply-To" header to the message ID of the cover letter
of the preceding version. This leads to "deep" nesting, where the nth
version of a series will be nested n-1 levels below the first version.

But not all projects prefer deep nesting. The alternative that is used
in some projects is "shallow" nesting: instead of nesting the rerolled
version under the preceding version's cover letter, it gets nested under
the first version's cover letter.

Implement support for a new "shallow" value for "b4.send-same-thread"
that does exactly this.

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Link: https://patch.msgid.link/20241112-pks-shallow-threads-v1-1-79466a503075@pks.im
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>