b4: stop generating full object IDs for non-binary diffs
The patches that are to be sent by b4 get generated via git-show(1).
These patches get generated with `--full-index`, which causes Git to
print the full object ID for the pre- and postimage blobs. This is
somewhat annoying to read and doesn't help in the general case:

  - It does not help in case the recipient doesn't have the necessary
    objects anyway.

  - It is unnecessarily long in order to protect from using the wrong
    object in a fake ancestor tree during a three-way merge.

The use of `--full-index` was added via 23a9ddb (Create patches with
--full-index --binary, 2024-08-27), which had the intent to include
binary diffs in generated patches. As part of that change we started to
pass both `--full-index` and `--binary`. Combining these two flags with
each other is not needed though: passing `--binary` already makes Git
print full object IDs for any binary patches, while keeping abbreviated
object IDs for normal patches.

Drop `--full-index` so that we print short object IDs in general, but
long object IDs for binaries.

Suggested-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Patrick Steinhardt <ps@pks.im>
Link: https://patch.msgid.link/20250115-b4-pks-no-full-index-v1-1-23a4e8fa83da@pks.im
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>