Partially revert "Don't invoke mailmap by default when sending patches"
This partially reverts commit 8438e306d91aca00bdc6dfedc182627281a2910a,
restricting the mailmap handling changes only to the git-show operation.
The consensus is that we *do* want to apply mailmap updates to the
actual addresses going into to/cc, just not to the in-body From:
addresses generated during the "git show" stage.

Link: https://lore.kernel.org/tools/e6c6f928-bae7-4509-9410-42e4bf7d9252@kernel.org/
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>