Fix trailer attribution bug due to incorrect thread traversing
When we were walking the thread looking for where the trailer belongs,
we were incorrectly attributing trailers sent to a single patch as if it
were sent to the cover letter. This only happened in the following
situation

 [PATCH 0/X] Cover Letter
 |
 |---> [PATCH Y/X] Patch Y
 |     (patch being reviewed)
 |
 |-----> Re: [PATCH Y/X] Patch Y
 |       (follow-up that *doesn't* contain a trailer)
 |
 |-------> Re: [PATCH Y/X] Patch Y
 |         (follow-up that *does* contains a trailer)

In this case we had a bug in traversing the references where we first
looked at the immediate parent, and then straight at the top of the
thread, becaues the references were not reversed when we had parsed
them.

This patch fixes both the incorrect order of references AND it doesn't
trust them to be in the correct order -- it will now always rely on the
in-reply-to header to figure out what the parent of the message is.

Reported-by: Andrew Cooper <andrew.cooper3@citrix.com>
Reported-by: Geert Uytterhoeven <geert@linux-m68k.org>
Fixes: https://bugzilla.kernel.org/show_bug.cgi?id=219993
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>