ez: introduce --in-reply-to for send
When sending numerous patchsets touching multiple projects, the
structure of the email thread could be top-level cover letter with
patchsets as subthreads (replies). As an example, while developing
a filesystem feature one can also develop tests for xfstests for
that feature.

With --in-reply-to one can send top-level cover and use it as base
for sending other b4 managed branches.

This patch introduces 'b4 --in-reply-to IDENTIFIER' for adding
'In-Reply-To:' header to cover-letter message.

Signed-off-by: Andrey Albershteyn <aalbersh@kernel.org>