ty: Allow users to specify whether thank-yous are sent to themselves
The motivation behind this change was that thank-yous weren't showing up
in maintainer's inboxes.  This is a problem because some of us the
presence of these mails to indicate that particular sets or individual
patches have been applied or not.

Provide command line option '--me-too' to override the default
behaviour.  If this is not specified, there are no functional changes.

Closes: https://bugzilla.kernel.org/show_bug.cgi?id=217672
Signed-off-by: Lee Jones <lee@kernel.org>
Link: https://patch.msgid.link/20250613133209.892846-1-lee@kernel.org
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>