Don't parse nonstandard git-version suffix
In snapshot builds of git, git-version yields a nonstandard version
format like 2.51.0.178.g2462961280. packaging.version.parse() errors out
on this with an InvalidVersion exception.

Since we only need to compare major and minor version, and
packaging.version is intended for a different version scheme, resort to
comparing the two numbers manually.

Fixes: 2e5aec1cd7ec ("Add Git version check to prevent crash with '--default-prefix' parameter")
Signed-off-by: Jonas Rebmann <jre@pengutronix.de>
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>