mbox: refactor how we treat base_commit for am and shazam
When using `b4 shazam` without -H or -M options, we don't really care to
know the base_commit, because we're applying directly on top of the
current head in all cases. Move all base_commit guessing logic into
parts of the code where we actually care to know what it is.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>