ez: cleanup sparse-checkout usage
Similarly to the previus patch, b4 ez uses sparse-checkout incorrectly:
- it does not want any files but it uses "init" which
  actually gives some files (the ones in the root)
- it runs checkout but it's pointless - index is left dirty,
  and it only happens to work because cherry-pick works
  with a dirty index (unlike e.g. am).

To fix, switch to sparse-checkout set to get an empty cone +
checkout -f to update the index.

We could actually skip checkout completely but it seems better
not to rely on the quirk of cherry-pick that allows it to
work with a dirty index, and an empty index is quick to update.

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>