Fix HEAD lookup checks for msgid existence
For a while now, HEAD requests have always been returning 302 or 301
redirects and we weren't actually bailing out early on non-existent
msgids. It still did the right thing at the actual mbox retrieval stage,
so it never really broke, but it will now bail out at the HEAD lookup
stage when the message-id is not known to the remote server.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>