Add support for setting a presubject
Some mailing lists require a prefix to come before the [PATCH] in the
subject of emails, such as OpenEmbedded or libgpiod. The current prefix
mechanism does not allow for that (only after PATCH), and setting the
format.subjectPrefix Git option does not work either.

Add a new configuration option send-presubject as well as a
--set-presubject argument to be able to include anything that should
come before the subject. The logic behind prefixes is not changed and
this new option does not break any existing prefix assignments.

In order not to break the logic in LoreSubject's __init__ function,
which parses the subject line and strips brackets in order to reorganize
the subject line later, the presubject must be enclosed with brackets
too (_check_presubject() is used to ensure that).

Closes: https://bugzilla.kernel.org/show_bug.cgi?id=217642
Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>